Pre-merge 2 - Missing : Navigator issue to fix + Refactor and use of AuthenticationModule + Notifications (waiting for final push on notification branch).
