+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:partnership/coordinator/COORDINATOR.dart';
+import 'package:partnership/coordinator/Coordinator.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 void main() {
